// libs/features/benefits/src/lib/shared/models/search-personal-information.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
import { PersonalInformation } from './personal-information';

/**
 * The wrapper class for search personal information.
 *
 * @export
 * @class SearchPersonalInformation
 */
export class SearchPersonalInformation {
  recordCount: number;
  listOfPersons: PersonalInformation[];
}